+++ src/Plainion.GraphViz.Viewer/CoreModule.cs
@@ -6,7 +6,7 @@ using Prism.Regions;
 
 namespace Plainion.GraphViz.Viewer
 {
-    [ModuleExport( typeof( CoreModule ) )]
+    [ModuleExport(typeof(CoreModule))]
     class CoreModule : IModule
     {
         private IRegionManager myRegionManager;
@@ -19,9 +19,9 @@ namespace Plainion.GraphViz.Viewer
 
         public void Initialize()
         {
-            myRegionManager.RegisterViewWithRegion( GraphViz.Viewer.RegionNames.GraphViewer, typeof( GraphViewer ) );
-            myRegionManager.RegisterViewWithRegion( GraphViz.Viewer.RegionNames.SettingsEditor, typeof( SettingsEditor ) );
-            myRegionManager.RegisterViewWithRegion( GraphViz.Viewer.RegionNames.StatusMessagesViewer, typeof( StatusMessagesView ) );
+            myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.GraphViewer, typeof(GraphViewer));
+            myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.SettingsEditor, typeof(SettingsEditor));
+            myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.StatusMessagesViewer, typeof(StatusMessagesView));
         }
     }
 }